# static/swagger.json
{
    "swagger": "2.0",
    "info": {
      "title": "Swagger App",
      "version": "6.6.6"
    },
    "paths": {
      "/api/electro-scooters": {
        "post": {
          "description": "Post a scooter",
          "consumes": [
            "application/json"
          ],
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "name": "Scooter",
              "in": "body",
              "description": "JSON data for the scooter",
              "required": true,
              "schema": {
                "$ref": "#/definitions/Scooter"
              }
            }
          ],
          "responses": {
            "201": {
              "description": "Scooter creation successful!"
            },
            "400": {
              "description": "Bad request"
            }
          }
        }
      },
    "/api/electro-scooters/{id}": {
        "get": {
          "description": "Returns a scooter by id",
          "consumes": [
            "id"
          ],
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "description": "Scooter id",
              "required": true,
              "schema": {
              "type": "integer",
              "minimum": 1
            }
            }
          ],
          "responses": {
            "200": {
              "description": "Success! Returned a scooter"
            },
            "400": {
              "description": "Bad request"
            }
          }
        },
        "put": {
          "description": "Update a scooter",
          "consumes": [
            "application/json"
          ],
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "description": "Scooter id",
              "required": true,
              "schema": {
              "type": "integer",
              "minimum": 1
            }
            },
            {
              "name": "Scooter",
              "in": "body",
              "description": "JSON data for the scooter",
              "required": true,
              "schema": {
                "$ref": "#/definitions/Scooter"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "Success! Scooter updated"
            },
            "400": {
              "description": "Bad request"
            }
          }
        },
        "delete": {
          "description": "Delete a scooter",
          "consumes": [
            "application/json"
          ],
          "produces": [
            "application/json"
          ],
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "description": "Scooter id",
              "required": true,
              "schema": {
              "type": "integer",
              "minimum": 1
            }
            },
            {
              "name": "X-Delete-Password",
              "in": "header",
              "description": "Password",
              "required": true,
              "schema": {
              "type": "string"
            }
            }
          ],
          "responses": {
            "200": {
              "description": "Success! Scooter deleted"
            },
            "400": {
              "description": "Bad request"
            }
          }
        }
      }
    },
    "definitions": {
      "Scooter": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "battery_level": {
            "type": "integer"
          }
        }
      }
    }
  }